Add files via upload
-Added `multiprocessing` to `Hedabot`. 
A single process deals with the sending and recieving of messages. A message recived will be send to a specific inputSorter processes based on the modulus of the `chat_id`. These processes then deal with the stack section of `Hedabot` and if a command has to be executed , sends the relevant information to a `CommandExecutor` process which then executes the command.
Default values are 4 inputSorters processes and 2 Execution processes per inputSorter. An increase in inputSorters may be benifitial if DB is slow or to reduce Execution process number to reduce memory usage as they are not often used.